# openapi.yml
openapi: 3.0.1
info:
  title: mitum-currency - digest API
  description: >-
    This is API description of digest service of mitum-currency.
  contact:
    name: Spike^ekipS
    url: https://github.com/spikeekips/mitum-currency
    email: spikeekips@gmail.com
  license:
    name: GPL-3.0
    url: https://github.com/spikeekips/mitum-currency/blob/master/LICENSE
  version: 0.0.1
...
servers:
- url: https://localhost:54322
...
paths:
...
  /block/operation/{operation_fact_hash}:
    get:
      tags:
      - block
      summary: Operation by it's fact hash.
      description: >-
        operation by it's **fact** *hash*, not it's own *hash*.
      operationId: operation
      parameters:
        - name: operation_fact_hash
          in: path
          description: >-
              *fact* *hash* of operation.
          required: true
          schema:
            type: string
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        404:
          description: not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of operation
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
  /block/manifests:
    get:
      tags:
      - block
      summary: All the manifest
      operationId: manifests
      parameters:
        - name: offset
          in: query
          schema:
            type: string
            example: "2"
          description: >-
            *manifest*s after *offset*, usually it is block *height*.
        - name: reverse
          in: query
          schema:
            type: boolean
            example: false
            default: false
          description: >-
            *manifest*s by reverse order.
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "...."
                      detail:
                        type: string
                        example: "...."
        404:
          description: no more manifests
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "manifests not found"
                      detail:
                        type: string
                        example: "...."
        200:
          description: hal document of manifests
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ManifestsHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
...
  /block/operations:
    get:
      tags:
      - block
      summary: All the operations
      operationId: operations
      parameters:
        - name: offset
          in: query
          schema:
            type: string
            example: "2,0"
          description: >-
            *operation*s after *offset*.
        - name: reverse
          in: query
          schema:
            type: boolean
            example: false
            default: false
          description: >-
            *operation*s by reverse order.
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "...."
                      detail:
                        type: string
                        example: "...."
        404:
          description: no more operations
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "operations not found"
                      detail:
                        type: string
                        example: "...."
        200:
          description: hal document of operations
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationsHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
  /account/{address}:
    get:
      tags:
      - account
      summary: The latest state of account
      description: >-
        The latest state of account. It contains the *keys* of account and it's *balance*.
      operationId: account
      parameters:
        - name: address
          in: path
          description: >
            *address* of account.
          required: true
          schema:
            $ref: '#/components/schemas/AccountAddress'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        404:
          description: not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of account
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/AccountHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
...
  /builder/operation:
    get:
      tags:
      - builder
      summary: Help to build operation messages.
      operationId: operation-builder
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of operation
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationBuilderHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
...
  /builder/operation/fact:
    post:
      tags:
      - builder
      summary: Help to build operation messages.
      description: >-
        It receives the edited fact message, not template HAL itself.
      operationId: operation-fact-builder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountsFact'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        400:
          description: problems in request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of operation template with signed fact.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationTemplateCreateAccountsHAL'
  /builder/operation/sign:
    post:
      tags:
      - builder
      summary: Help to sign operation messages.
      description: >-
        It receives the edited operation message, not template HAL itself.
      operationId: operation-fact-builder-sign
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccounts'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        400:
          description: problems in request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of operation template with signed fact.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationTemplateCreateAccountsSignHAL'
  /builder/send:
    post:
      tags:
      - builder
      summary: Broadcast message to network
      description: >-
        It receives the operation or seal and broadcasts to the network. The signer of seal(not *operation*) will be digest node itself. Such as,

        - If request body is *operation*, new seal will be created with the given *operation* and the new seal will be signed by digest node.

        - If request body is *seal*, it will be signed by digest node.
      operationId: operation-builder-send
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Seal'
                - $ref: '#/components/schemas/CreateAccounts'
                - $ref: '#/components/schemas/KeyUpdater'
                - $ref: '#/components/schemas/Transfers'
                - $ref: '#/components/schemas/CurrencyRegister'
                - $ref: '#/components/schemas/CurrencyPolicyUpdater'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        400:
          description: problems in request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of operation template with signed fact.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationTemplateCreateAccountsHAL'
  /currency:
    get:
      tags:
      - currency
      summary: Currencies information
      operationId: currencies
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of currencies
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CurrenciesHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
  /currency/{currency_id}:
    get:
      tags:
      - currency
      summary: Currency information
      operationId: currency
      parameters:
        - name: currency_id
          in: path
          description: currency unique id(or name)
          required: true
          schema:
            $ref: '#/components/schemas/CurrencyID'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of currency of *currency_id*
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CurrencyHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
components:
  schemas:
...
    Problem:
      type: object
      required:
      - _hint
      - title
      - type
      - detail
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: a017:0.0.1
              default: a017:0.0.1
        title:
          type: string
          example: "failed to sign"
        type:
          type: string
          example: "https://github.com/spikeekips/mitum-currency/problems/others"
        detail:
          type: string
          example: "invalid privatekey"
...
    AccountHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-account-value
                  example: mitum-currency-account-value
                hint:
                  type: string
                  default: a018:0.0.1
                  example: a018:0.0.1
            _embedded:
              $ref: '#/components/schemas/AccountValue'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1
                operations:
                  description: >-
                    *operation*s, which are related of the account.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations
                operations:{offset}:
                  description: >-
                    *operation*s, which are related of the account after offset.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations?offset={offset}
                        templated:
                          type: boolean
                          default: true
                          example: true
                operations:{offset,reverse}:
                  description: >-
                    *operation*s, which are related of the account after offset by reverse order.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations?offset={offset&reverse=1}
                        templated:
                          type: boolean
                          default: true
                          example: true
                block:
                  description: >-
                    Request `/block/{height}`.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244
...
    OperationHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-operation-value
                  example: mitum-currency-operation-value
                hint:
                  type: string
                  default: a019:0.0.1
                  example: a019:0.0.1
            _embedded:
              $ref: '#/components/schemas/OperationValue'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /operation/6GymxmuvhgUfAKWYQUKKveYh1H8b981vQaZcMsbWqykS
                operation:{hash}:
                  description: request `/operation/{hash}`
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          default: /operation/{hash:(?i)[0-9a-z][0-9a-z]+}
                          example: /operation/{hash:(?i)[0-9a-z][0-9a-z]+}
                        templated:
                          type: boolean
                          default: true
                          example: true
                block:
                  description: request `/block/{height}` of the operation
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244
                manifest:
                  description: request `/block/{height}/manifest` of the operation
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244/manifest
...
    ManifestsHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            _embedded:
              type: array
              items:
                $ref: '#/components/schemas/ManifestHAL'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/manifests
                next:
                  description: >-
                    next manifests with *offset*.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/manifests?offset=2
                reverse:
                  description: >-
                    manifests by reverse oder of self.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/manifests?reverse=1
    OperationsHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            _embedded:
              type: array
              items:
                $ref: '#/components/schemas/OperationHAL'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations
                next:
                  description: >-
                    next operations with *offset*.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations?offset=2,0
                reverse:
                  description: >-
                    operations by reverse oder of self.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations?reverse=1
...
    CreateAccounts:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/CreateAccountsFact'
    KeyUpdater:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/KeyUpdaterFact'
    Transfers:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/TransfersFact'
...
    CreateAccountsFact:
      allOf:
        - $ref: '#/components/schemas/BaseFact'
        - type: object
          required:
          - sender
          - items
          properties:
            _hint:
              allOf:
                - $ref: '#/components/schemas/Hint'
                - type: string
                  example: a005:0.0.1
                  default: a005:0.0.1
            hash:
              description: >-
                The value of hash will be generated automatically by builder. *Don't need to edit*.
              type: string
              format: hash
              example: 4jhzcudKgtoPGR6rA7Fuxmfwz3C8KGiP5MEKXuBXcW9j
            token:
              description: >-
                Replace your own token. *token* value should be encoded by base64.
              type: string
              format: bytes
              example: cmFpc2VkIGJ5
            sender:
              allOf:
                - $ref: '#/components/schemas/AccountAddress'
                - description: Replace your own sender address.
            items:
              type: array
              items:
                type: object
                required:
                - keys
                - amount
                properties:
                  keys:
                    $ref: '#/components/schemas/AccountKeys'
                  amounts:
                    type: array
                    items:
                      description: The initial balance of account.
                      allOf:
                        - $ref: '#/components/schemas/Amount'
...
    CurrencyID:
      description: currency unique id(or name)
      type: string
      example: XXX
...
    AccountAddress:
      description: account address
      type: string
      example: 'mc-node-010a:0.0.1'

# vi: ft=yaml tw=100 ts=2 sw=2 expandtab smarttab